Extract configuration and profiles into a configuration package.
