+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/configuration/AbstractSecurityConfig.java
@@ -1,4 +1,4 @@
-package org.optaweb.employeerostering;
+package org.optaweb.employeerostering.configuration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,10 +7,6 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
-class Profiles {
-    public static final String DEVELOPMENT = "dev";
-    public static final String PRODUCTION = "prod";
-}
 
 abstract public class AbstractSecurityConfig extends WebSecurityConfigurerAdapter {}
