+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/configuration/Profiles.java
@@ -1,0 +1,7 @@
+package org.optaweb.employeerostering.configuration;
+
+public class Profiles {
+    public static final String DEVELOPMENT = "dev";
+    public static final String PRODUCTION = "prod";
+    public static final String DEFAULT = "default";
+}
